files compression

+++ src/modules/users/users.service.ts
@@ -1,4 +1,5 @@
-import { Injectable } from '@nestjs/common';
+import { BadRequestException, Injectable, NotFoundException } from '@nestjs/common';
+import { isUUID } from 'class-validator';
 import { PrismaService } from 'src/prisma/prisma.service';
 
 @Injectable()
@@ -6,16 +7,19 @@ export class UserService {
   constructor(private prisma: PrismaService) {}
 
   async findAll() {
-    const users = await this.prisma.user.findMany({
-      include: {
-        campings: true,
-      },
-    });
+
+    const users = await this.prisma.user.findMany();
+
+    if (!users) throw new NotFoundException('Users not found');
+
 
     return users.map(({ password, ...userWithoutPassword }) => userWithoutPassword);
   }
 
   async findById(id: string) {
+
+    
+    
     const user = await this.prisma.user.findUnique({
       where: {
         id,
@@ -24,14 +28,27 @@ export class UserService {
         campings: true,
       },
     });
+    
+    if (!user) throw new NotFoundException('User not found');
+    
+    if (!isUUID(id)) throw new BadRequestException('User ID not valid');
+
 
     const { password, ...userWithoutPassword } = user;
     return userWithoutPassword;
   }
 
-  async create(data: any) {}
+
+  // async create(data: any) {}
 
   async update(id: string, data: any) {
+
+    if (!isUUID(id)) throw new BadRequestException('User ID not valid');
+
+    const user = await this.prisma.user.findUnique({where: {id}})
+
+    if (!user) throw new NotFoundException('User not found');
+
     return await this.prisma.user.update({
       where: {
         id,
@@ -41,11 +58,16 @@ export class UserService {
   }
 
   async delete(id: string) {
+
+    if (!isUUID(id)) throw new BadRequestException('User ID not valid');
+
     const user = await this.prisma.user.delete({
       where: {
         id,
       },
     });
+    
+    if (!user) throw new NotFoundException('User not found');
 
     const { password, ...userWithoutPassword } = user;
     return userWithoutPassword;
